requirements.txt
streamlit
python-docx
PyMuPDF
ollama
